+++ core/shared/src/main/scala/zio/logging/package.scala
@@ -15,13 +15,7 @@
  */
 package zio
 
-import zio.metrics.{ Metric, MetricLabel }
-
-import java.io.PrintStream
-import java.nio.charset.{ Charset, StandardCharsets }
-import java.nio.file.Path
-
-package object logging {
+package object logging extends LoggerLayers {
 
   /**
    * The [[logContext]] fiber reference is used to store typed, structured log
@@ -50,10 +44,6 @@ package object logging {
    */
   val loggerNameAnnotationKey = "logger_name"
 
-  private[logging] val logLevelMetricLabel = "level"
-
-  private[logging] val loggedTotalMetric = Metric.counter(name = "zio_log_total")
-
   /**
    * Logger name aspect, by this aspect is possible to set logger name (in general, logger name is extracted from [[Trace]])
    *
@@ -61,268 +51,6 @@ package object logging {
    */
   def loggerName(value: String): ZIOAspect[Nothing, Any, Nothing, Any, Nothing, Any] =
     ZIOAspect.annotated(loggerNameAnnotationKey, value)
-
-  def consoleErrLogger(config: ConsoleLoggerConfig): ZLayer[Any, Nothing, Unit] =
-    Runtime.addLogger(makeConsoleErrLogger(config))
-
-  def consoleErrJsonLogger(config: ConsoleLoggerConfig): ZLayer[Any, Nothing, Unit] =
-    Runtime.addLogger(makeConsoleErrJsonLogger(config))
-
-  def consoleErrJsonLogger(configPath: String = "logger"): ZLayer[Any, Config.Error, Unit] =
-    ZLayer.scoped {
-      for {
-        config <- ConsoleLoggerConfig.load(configPath)
-        _      <- ZIO.withLoggerScoped(makeConsoleErrJsonLogger(config))
-      } yield ()
-    }
-
-  def consoleErrLogger(configPath: String = "logger"): ZLayer[Any, Config.Error, Unit] =
-    ZLayer.scoped {
-      for {
-        config <- ConsoleLoggerConfig.load(configPath)
-        _      <- ZIO.withLoggerScoped(makeConsoleErrLogger(config))
-      } yield ()
-    }
-
-  def consoleJsonLogger(config: ConsoleLoggerConfig): ZLayer[Any, Nothing, Unit] =
-    Runtime.addLogger(makeConsoleJsonLogger(config))
-
-  def consoleJsonLogger(configPath: String = "logger"): ZLayer[Any, Config.Error, Unit] =
-    ZLayer.scoped {
-      for {
-        config <- ConsoleLoggerConfig.load(configPath)
-        _      <- ZIO.withLoggerScoped(makeConsoleJsonLogger(config))
-      } yield ()
-    }
-
-  def consoleLogger(config: ConsoleLoggerConfig): ZLayer[Any, Nothing, Unit] =
-    Runtime.addLogger(makeConsoleLogger(config))
-
-  def consoleLogger(configPath: String = "logger"): ZLayer[Any, Config.Error, Unit] =
-    ZLayer.scoped {
-      for {
-        config <- ConsoleLoggerConfig.load(configPath)
-        _      <- ZIO.withLoggerScoped(makeConsoleLogger(config))
-      } yield ()
-    }
-
-  def fileAsyncJsonLogger(config: FileLoggerConfig): ZLayer[Any, Nothing, Unit] =
-    ZLayer.scoped(makeFileAsyncJsonLogger(config))
-
-  def fileAsyncJsonLogger(configPath: String = "logger"): ZLayer[Any, Config.Error, Unit] =
-    ZLayer.scoped {
-      for {
-        config <- FileLoggerConfig.load(configPath)
-        _      <- makeFileAsyncJsonLogger(config)
-      } yield ()
-    }
-
-  def fileAsyncLogger(config: FileLoggerConfig): ZLayer[Any, Nothing, Unit] =
-    ZLayer.scoped(makeFileAsyncLogger(config))
-
-  def fileAsyncLogger(configPath: String = "logger"): ZLayer[Any, Config.Error, Unit] =
-    ZLayer.scoped {
-      for {
-        config <- FileLoggerConfig.load(configPath)
-        _      <- makeFileAsyncLogger(config)
-      } yield ()
-    }
-
-  def fileJsonLogger(config: FileLoggerConfig): ZLayer[Any, Nothing, Unit] =
-    Runtime.addLogger(makeFileJsonLogger(config))
-
-  def fileJsonLogger(configPath: String = "logger"): ZLayer[Any, Config.Error, Unit] =
-    ZLayer.scoped {
-      for {
-        config <- FileLoggerConfig.load(configPath)
-        _      <- ZIO.withLoggerScoped(makeFileJsonLogger(config))
-      } yield ()
-    }
-
-  def fileLogger(config: FileLoggerConfig): ZLayer[Any, Nothing, Unit] =
-    Runtime.addLogger(makeFileLogger(config))
-
-  def fileLogger(configPath: String = "logger"): ZLayer[Any, Config.Error, Unit] =
-    ZLayer.scoped {
-      for {
-        config <- FileLoggerConfig.load(configPath)
-        _      <- ZIO.withLoggerScoped(makeFileLogger(config))
-      } yield ()
-    }
-
-  val logMetrics: ZLayer[Any, Nothing, Unit] =
-    Runtime.addLogger(makeMetricLogger(loggedTotalMetric, logLevelMetricLabel))
-
-  def logMetricsWith(name: String, logLevelLabel: String): ZLayer[Any, Nothing, Unit] =
-    Runtime.addLogger(makeMetricLogger(Metric.counter(name), logLevelLabel))
-
-  private def makeConsoleErrLogger(config: ConsoleLoggerConfig): ZLogger[String, Any] =
-    makeConsoleLogger(config.format.toLogger, java.lang.System.err, config.filter)
-
-  private def makeConsoleErrJsonLogger(config: ConsoleLoggerConfig): ZLogger[String, Any] =
-    makeConsoleLogger(config.format.toJsonLogger, java.lang.System.err, config.filter)
-
-  private def makeConsoleLogger(config: ConsoleLoggerConfig): ZLogger[String, Any] =
-    makeConsoleLogger(config.format.toLogger, java.lang.System.out, config.filter)
-
-  private def makeConsoleJsonLogger(config: ConsoleLoggerConfig): ZLogger[String, Any] =
-    makeConsoleLogger(config.format.toJsonLogger, java.lang.System.out, config.filter)
-
-  private def makeConsoleLogger(
-    logger: ZLogger[String, String],
-    stream: PrintStream,
-    logFilter: LogFilter[String]
-  ): ZLogger[String, Any] = {
-
-    val stringLogger = logFilter.filter(logger.map { line =>
-      try stream.println(line)
-      catch {
-        case t: VirtualMachineError => throw t
-        case _: Throwable           => ()
-      }
-    })
-    stringLogger
-  }
-
-  private def makeFileAsyncJsonLogger(
-    config: FileLoggerConfig
-  ): ZIO[Scope, Nothing, Unit] = makeFileAsyncLogger(
-    config.destination,
-    config.format.toJsonLogger,
-    config.filter,
-    config.charset,
-    config.autoFlushBatchSize,
-    config.bufferedIOSize,
-    config.rollingPolicy
-  )
-
-  private def makeFileAsyncLogger(
-    config: FileLoggerConfig
-  ): ZIO[Scope, Nothing, Unit] = makeFileAsyncLogger(
-    config.destination,
-    config.format.toLogger,
-    config.filter,
-    config.charset,
-    config.autoFlushBatchSize,
-    config.bufferedIOSize,
-    config.rollingPolicy
-  )
-
-  private def makeFileAsyncLogger(
-    destination: Path,
-    logger: ZLogger[String, String],
-    logFilter: LogFilter[String],
-    charset: Charset,
-    autoFlushBatchSize: Int,
-    bufferedIOSize: Option[Int],
-    rollingPolicy: Option[FileLoggerConfig.FileRollingPolicy]
-  ): ZIO[Scope, Nothing, Unit] =
-    for {
-      queue       <- Queue.bounded[UIO[Any]](1000)
-      stringLogger =
-        makeFileAsyncLogger(
-          destination,
-          logger,
-          logFilter,
-          charset,
-          autoFlushBatchSize,
-          bufferedIOSize,
-          queue,
-          rollingPolicy
-        )
-      _           <- ZIO.withLoggerScoped(stringLogger)
-      _           <- queue.take.flatMap(task => task.ignore).forever.forkScoped
-    } yield ()
-
-  private def makeFileAsyncLogger(
-    destination: Path,
-    logger: ZLogger[String, String],
-    logFilter: LogFilter[String],
-    charset: Charset,
-    autoFlushBatchSize: Int,
-    bufferedIOSize: Option[Int],
-    queue: Queue[UIO[Any]],
-    rollingPolicy: Option[FileLoggerConfig.FileRollingPolicy]
-  ): ZLogger[String, Any] = {
-    val logWriter =
-      new zio.logging.internal.FileWriter(destination, charset, autoFlushBatchSize, bufferedIOSize, rollingPolicy)
-
-    val stringLogger: ZLogger[String, Any] = logFilter.filter(logger.map { (line: String) =>
-      zio.Unsafe.unsafe { implicit u =>
-        Runtime.default.unsafe.run(queue.offer(ZIO.succeed {
-          try logWriter.writeln(line)
-          catch {
-            case t: VirtualMachineError => throw t
-            case _: Throwable           => ()
-          }
-        }))
-      }
-    })
-    stringLogger
-  }
-
-  private def makeFileJsonLogger(config: FileLoggerConfig): ZLogger[String, Any] =
-    makeFileLogger(
-      config.destination,
-      config.format.toJsonLogger,
-      config.filter,
-      config.charset,
-      config.autoFlushBatchSize,
-      config.bufferedIOSize,
-      config.rollingPolicy
-    )
-
-  private def makeFileLogger(config: FileLoggerConfig): ZLogger[String, Any] =
-    makeFileLogger(
-      config.destination,
-      config.format.toLogger,
-      config.filter,
-      config.charset,
-      config.autoFlushBatchSize,
-      config.bufferedIOSize,
-      config.rollingPolicy
-    )
-
-  private def makeFileLogger(
-    destination: Path,
-    logger: ZLogger[String, String],
-    logFilter: LogFilter[String],
-    charset: Charset,
-    autoFlushBatchSize: Int,
-    bufferedIOSize: Option[Int],
-    rollingPolicy: Option[FileLoggerConfig.FileRollingPolicy]
-  ): ZLogger[String, Any] = {
-    val logWriter =
-      new zio.logging.internal.FileWriter(destination, charset, autoFlushBatchSize, bufferedIOSize, rollingPolicy)
-
-    val stringLogger: ZLogger[String, Any] = logFilter.filter(logger.map { (line: String) =>
-      try logWriter.writeln(line)
-      catch {
-        case t: VirtualMachineError => throw t
-        case _: Throwable           => ()
-      }
-    })
-
-    stringLogger
-  }
-
-  private def makeMetricLogger(counter: Metric.Counter[Long], logLevelLabel: String): ZLogger[String, Unit] =
-    new ZLogger[String, Unit] {
-      override def apply(
-        trace: Trace,
-        fiberId: FiberId,
-        logLevel: LogLevel,
-        message: () => String,
-        cause: Cause[Any],
-        context: FiberRefs,
-        spans: List[LogSpan],
-        annotations: Map[String, String]
-      ): Unit = {
-        val tags = context.get(FiberRef.currentTags).getOrElse(Set.empty)
-        counter.unsafe.update(1, tags + MetricLabel(logLevelLabel, logLevel.label.toLowerCase))(Unsafe.unsafe)
-        ()
-      }
-    }
 
   val removeDefaultLoggers: ZLayer[Any, Nothing, Unit] = Runtime.removeDefaultLoggers
